Fix for modern node
This includes various fixes for modern npm (Which does not have the API from NPM 7 anymore), adds yarn support, moves create-probot-app to ESM, and requires Node 16.

Requiring Node 16 for create-probot-app should help encourage people to use a modern node for new apps, so migratting probot to require Node 14 or 16 in the future should be easier.